Changing field type to Doub;e

+++ src/main/java/com/locations/batch/locationsbatch/dto/LocationDTO.java
@@ -1,7 +1,5 @@
 package com.locations.batch.locationsbatch.dto;
 
-import java.math.BigDecimal;
-
 import org.apache.commons.lang3.math.NumberUtils;
 
 public class LocationDTO {
@@ -17,8 +15,8 @@ public class LocationDTO {
 	private String district;
 	private String street;
 	private Long streetNumber;
-	private BigDecimal latitude;
-	private BigDecimal longitude;
+	private Double latitude;
+	private Double longitude;
 	private String category;
 
 	public String getId() {
@@ -111,19 +109,19 @@ public class LocationDTO {
 		}
 	}
 
-	public BigDecimal getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(BigDecimal latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public BigDecimal getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(BigDecimal longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
